requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.6
et_xmlfile==2.0.0
exceptiongroup==1.3.1
idna==3.11
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.3
pyaes==1.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vk_api==11.10.0
